+++ CHANGELOG.md
@@ -17,6 +17,13 @@
   - .showRecordId(true) // Set the cursor showRecordId
   - .stream(options) // Return a modified Readable stream including a possible transform method.
 
+## [1.2.0] - 2018-10-13
+
+### Added
+
+- New deleteMany() method
+- Documentation for deleteMany() method
+
 ## [1.1.0] - 2018-09-10
 
 ### Fixed
@@ -571,6 +578,7 @@
 - findOne() method.
 - Document inheritence support.
 
+[1.2.0]: https://github.com/Yonirt/moltyjs/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Yonirt/moltyjs/compare/v1.0.10...v1.1.0
 [1.0.10]: https://github.com/Yonirt/moltyjs/compare/v1.0.9...v1.0.10
 [1.0.9]: https://github.com/Yonirt/moltyjs/compare/v1.0.8...v1.0.9